requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.10.0
python-dotenv>=1.0.0
websocket-client>=1.0.0
aiohttp>=3.8.0
colorama>=0.4.0
tabulate>=0.9.0
numpy>=1.24.0
